Simplify traits and mixins.
