+++ lib/ModelMixins/UrlMixin.ts
@@ -1,12 +1,10 @@
 import { computed } from 'mobx';
 import URI from 'urijs';
 import Constructor from '../Core/Constructor';
+import Model from '../Models/Model';
+import UrlTraits from '../Traits/UrlTraits';
 
-interface RequiredOnInstance {
-    url: string | undefined;
-}
-
-export default function UrlMixin<T extends Constructor<RequiredOnInstance>>(Base: T) {
+export default function UrlMixin<T extends Constructor<Model<UrlTraits>>>(Base: T) {
     class UrlMixin extends Base {
         @computed
         get uri(): uri.URI | undefined {
