+++ lib/Traits/UrlTraits.ts
@@ -1,0 +1,11 @@
+import ModelTraits from './ModelTraits';
+import primitiveTrait from './primitiveTrait';
+
+export default class UrlTraits extends ModelTraits {
+    @primitiveTrait({
+        type: 'string',
+        name: 'URL',
+        description: 'The base URL of the file or service.'
+    })
+    url?: string;
+}
