+++ test/Traits/objectArrayTraitSpec.ts
@@ -1,11 +1,10 @@
-import { configure, runInAction, autorun } from 'mobx';
-import primitiveTrait from '../../lib/Traits/primitiveTrait';
-import objectTrait from '../../lib/Traits/objectTrait';
-import ModelTraits from '../../lib/Traits/ModelTraits';
-import Model from '../../lib/Models/Model';
-import Terria from '../../lib/Models/Terria';
-import objectArrayTrait from '../../lib/Traits/objectArrayTrait';
+import { configure } from 'mobx';
+import CreateModel from '../../lib/Models/CreateModel';
 import createStratumInstance from '../../lib/Models/createStratumInstance';
+import Terria from '../../lib/Models/Terria';
+import ModelTraits from '../../lib/Traits/ModelTraits';
+import objectArrayTrait from '../../lib/Traits/objectArrayTrait';
+import primitiveTrait from '../../lib/Traits/primitiveTrait';
 
 configure({
     enforceActions: true,
@@ -49,7 +48,7 @@ class OuterTraits extends ModelTraits {
     inner?: InnerTraits[];
 }
 
-class TestModel extends Model(OuterTraits) {
+class TestModel extends CreateModel(OuterTraits) {
 
 }
